Finishing MVP stage

+++ Application/Posts/Delete.cs
@@ -1,5 +1,7 @@
 ﻿using Application.Core;
+using Application.Interfaces;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using Persistance;
 using System;
 using System.Collections.Generic;
@@ -20,16 +22,19 @@ namespace Application.Posts
         public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
+            private readonly IUserAccessor _userAccessor;
 
-            public Handler(DataContext context)
+            public Handler(DataContext context, IUserAccessor userAccessor)
             {
                 _context = context;
+                _userAccessor = userAccessor;
             }
             public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
-                var post = await _context.Posts.FindAsync(request.Id);
+                var post = await _context.Posts.Include(post => post.User).FirstOrDefaultAsync(x => x.Id == request.Id);
 
-                //if (post == null) return null;
+                if (_userAccessor.GetUsername() != post.User.UserName)
+                    return Result<Unit>.Failure("Nie udało się usunąć posta");
 
                 _context.Posts.Remove(post);
